requirements.txt
requests
pyyaml
curlify
